Player Movement Animation
1. Player Movement Animation 구현
2. 대각선 이동속도 조정(단위벡터 사용하여 조정)
3. L-Shift 키 눌렀을 시 대쉬하는 기능 추가

+++ SpriteSheet.cpp
@@ -40,7 +40,7 @@ void SpriteSheet::Draw()
 	);
 }
 
-void SpriteSheet::Draw(UINT index, UINT x, UINT y)
+void SpriteSheet::Draw(UINT index, FLOAT x, FLOAT y)
 {
 	D2D1_RECT_F src = D2D1::RectF(
 		(FLOAT)((index % m_spriteAcross) * m_spriteWidth),
